feat: refactor WithdrawModal and integrate it into TokenHoldingsTable for improved user experience
